requirements.txt
flask>=3.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
unidecode>=1.3.0
werkzeug>=3.0.0
